test: 차, 포, 졸 궁 내 대각선 이동 테스트 작성
